Split hardware_service subsystems into services/* packages
Phase 2 of the hardware_service.py split.  Lifts the per-subsystem
init helpers, the ZigpyController class, the GPS / chrony trend
archive, and the cross-subsystem metrics publisher out of the
3,165-line monolith and into four new packages under services/.

Per-subsystem packages
----------------------
* services/gps/
  - trends.py   : RRD-style sampler (raw → 1m → 10m → 1h tiers) as
                  pure library functions that accept the redis client
                  and per-tier last-bucket-id dict as arguments.
  - manager.py  : initialize_gps_manager() — returns the started
                  GPSManager so the orchestrator owns lifetime.
* services/zigbee/
  - detection.py  : detect_zigbee_coordinator() + USB signatures.
  - controller.py : ZigpyController (asyncio-driven zigpy-znp).
  - coordinator.py: initialize_zigbee_coordinator() +
                    publish_zigbee_status().
* services/displays/
  - init.py  : LED / VFD / OLED / screen-manager bootstrap.
  - state.py : publish_display_state() with preview-image capture.
* services/gpio/
  - init.py             : GPIO / NeoPixel / USB tower-light bootstrap.
  - alert_indicators.py : tower-light + NeoPixel state machine driven
                          by eas:broadcast_active / eas:incoming_alert.
* services/common/metrics.py — publish_hardware_metrics() coordinates
  the display + zigbee publishers and the high-level hardware:metrics
  key (kept under common/ because it spans subsystems).

Orchestrator
------------
hardware_service.py shrinks from 3,165 → 2,007 lines.  All module
globals (_redis_client, _flask_app, _screen_manager, _gpio_controller,
_neopixel_controller, _tower_light_controller, _gps_manager,
_zigpy_controller, _gps_trend_last_bucket_ids) stay here so the Flask
API and shutdown handler still see the live state; the legacy public
names (initialize_*, publish_*, _aggregate_gps_trend_samples,
_emit_gps_trend_rollups, ZigpyController, detect_zigbee_coordinator,
GPS_TRENDS_* constants) are preserved as thin wrappers that thread
the orchestrator state into the new pure modules.

No behaviour change — Flask API endpoints (~1,500 lines, deferred to
Phase 3) are untouched, the systemd unit still launches one process,
and tests/test_gps_trends_archive.py keeps working unchanged thanks
to the re-export surface.
