+++ services/displays/state.py
@@ -1,0 +1,141 @@
+"""
+EAS Station - Emergency Alert System
+Copyright (c) 2025-2026 Timothy Kramer (KR8MER)
+
+This file is part of EAS Station.
+
+EAS Station is dual-licensed software:
+- GNU Affero General Public License v3 (AGPL-3.0) for open-source use
+- Commercial License for proprietary use
+
+You should have received a copy of both licenses with this software.
+For more information, see LICENSE and LICENSE-COMMERCIAL files.
+
+IMPORTANT: This software cannot be rebranded or have attribution removed.
+See NOTICE file for complete terms.
+
+Repository: https://github.com/KR8MER/eas-station
+"""
+
+from __future__ import annotations
+
+"""Publish detailed display state (with preview images) to Redis."""
+
+import json
+import logging
+from datetime import datetime, timezone
+
+logger = logging.getLogger(__name__)
+
+
+def publish_display_state(redis_client, screen_manager) -> None:
+    """Publish detailed display state including preview images to Redis."""
+    if not redis_client:
+        return
+
+    try:
+        # Import hardware settings helpers
+        from app_core.hardware_settings import get_oled_settings, get_led_settings, get_vfd_settings
+
+        state = {
+            "timestamp": datetime.now(timezone.utc).isoformat(),
+            "oled": {
+                "enabled": False,
+                "width": 128,
+                "height": 64,
+                "current_screen": None,
+                "scroll_offset": 0,
+                "alert_active": False,
+            },
+            "vfd": {
+                "enabled": False,
+                "width": 140,
+                "height": 32,
+                "current_screen": None,
+            },
+            "led": {
+                "enabled": False,
+                "lines": 4,
+                "chars_per_line": 20,
+                "current_message": None,
+                "color": "AMBER",
+            },
+        }
+
+        # Get OLED state from database and module
+        try:
+            oled_settings = get_oled_settings()
+            oled_enabled_in_db = oled_settings.get('enabled', False)
+
+            # Import after getting settings to avoid circular imports
+            import app_core.oled as oled_module
+
+            # Only show as enabled if both database setting is true AND controller exists
+            if oled_enabled_in_db and oled_module.oled_controller:
+                state["oled"]["enabled"] = True
+                state["oled"]["width"] = oled_module.oled_controller.width
+                state["oled"]["height"] = oled_module.oled_controller.height
+
+                # Get current screen name if available
+                if screen_manager and hasattr(screen_manager, '_current_oled_screen'):
+                    current_screen = screen_manager._current_oled_screen
+                    if current_screen:
+                        state["oled"]["current_screen"] = current_screen.name if hasattr(current_screen, 'name') else str(current_screen)
+
+                # Get current alert state if scrolling
+                if screen_manager:
+                    if hasattr(screen_manager, '_oled_scroll_effect') and screen_manager._oled_scroll_effect:
+                        state["oled"]["alert_active"] = True
+                        state["oled"]["scroll_offset"] = getattr(screen_manager, '_oled_scroll_offset', 0)
+                        state["oled"]["alert_text"] = getattr(screen_manager, '_current_alert_text', "") or ""
+                        state["oled"]["scroll_speed"] = getattr(screen_manager, '_oled_scroll_speed', 4)
+
+                        # Get cached header
+                        if hasattr(screen_manager, '_cached_header_text'):
+                            state["oled"]["header_text"] = screen_manager._cached_header_text
+
+                # Get preview image
+                try:
+                    preview_image = oled_module.oled_controller.get_preview_image_base64()
+                    if preview_image:
+                        state["oled"]["preview_image"] = preview_image
+                except Exception as e:
+                    logger.debug(f"Failed to get OLED preview image: {e}")
+        except Exception as e:
+            logger.debug(f"Error getting OLED state: {e}")
+
+        # Get VFD state from database and module
+        try:
+            vfd_settings = get_vfd_settings()
+            vfd_enabled_in_db = vfd_settings.get('enabled', False)
+
+            from app_core.vfd import vfd_controller
+
+            # Only show as enabled if both database setting is true AND controller exists
+            if vfd_enabled_in_db and vfd_controller:
+                state["vfd"]["enabled"] = True
+        except Exception as e:
+            logger.debug(f"Error getting VFD state: {e}")
+
+        # Get LED state from database and module
+        try:
+            led_settings = get_led_settings()
+            led_enabled_in_db = led_settings.get('enabled', False)
+
+            import app_core.led as led_module
+
+            # Only show as enabled if both database setting is true AND controller exists
+            if led_enabled_in_db and led_module.led_controller:
+                state["led"]["enabled"] = True
+        except Exception as e:
+            logger.debug(f"Error getting LED state: {e}")
+
+        # Publish to Redis with short TTL (refreshes every 5 seconds)
+        redis_client.setex(
+            "hardware:display_state",
+            15,  # 15 second TTL (3x the publish interval for tolerance)
+            json.dumps(state)
+        )
+
+    except Exception as e:
+        logger.debug(f"Failed to publish display state: {e}")
